improve: components of product detail screen create
